resolve mypy bug
