+++ starlite/asgi/routing_trie/types.py
@@ -59,3 +59,22 @@ class RouteTrieNode(TypedDict):
     """
     A list of tuples containing path parameter definitions. This is used for parsing extracted path parameter values.
     """
+
+
+def create_node() -> "RouteTrieNode":
+    """Creates a RouteMapNode instance.
+
+    Returns:
+        A route map node instance.
+    """
+
+    return {
+        "asgi_handlers": {},
+        "child_keys": set(),
+        "children": {},
+        "is_asgi": False,
+        "is_mount": False,
+        "is_static": False,
+        "is_path_type": False,
+        "path_parameters": [],
+    }
